+++ routes/vehicleTypes.js
@@ -1,7 +1,7 @@
+// Require JavaScript Libraries
 const express = require('express');
 const router = express.Router();
 const { ensureAuthenticated, authenticateServer } = require('../helpers/auth');
-const mongoose = require('mongoose');
 const axios = require('axios');
 const keys = require('../config/keys');
 
@@ -19,7 +19,12 @@ router.get('/',
         companyID: req.session.companyID
       })
     })
-    .catch(error => console.log(error));
+    .catch(error => {
+      //console.log(error)
+      res.render('index/errorPage', {
+        error: error
+      }) 
+    });
 });
 
 // Vehicle Types - Add Page
@@ -39,17 +44,37 @@ router.get('/edit/:id',
   authenticateServer,
   ensureAuthenticated,
   (req, res) => {
-    axios.get(`${keys.sassTransferServiceAPIURI}/api/VehicleTypes/${req.params.id}?&access_token=${req.session.serverAccessToken}`)
-    .then(response => {
-      res.render('dashboard', {
-        dashboardLink: true,
-        vehicleTypesEditActive: true,
-        companyID: req.session.companyID,
-        vehicleType: response.data
-      })
+    axios.get(`${keys.sassTransferServiceAPIURI}/api/VehicleTypes/${req.params.id}?access_token=${req.session.serverAccessToken}`)
+      .then(response => {
+        // Check that the resouce belongs to the user requesting it
+        if (response.data.companyId != req.session.companyID) {
+          req.flash('error_msg', 'Unauthorized');
+          res.redirect('/dashboard');
+        }
+        else {
+          axios.get(`${keys.sassTransferServiceAPIURI}/api/VehicleTypes/${req.params.id}?&access_token=${req.session.serverAccessToken}`)
+            .then(response => {
+              res.render('dashboard', {
+                dashboardLink: true,
+                vehicleTypesEditActive: true,
+                companyID: req.session.companyID,
+                vehicleType: response.data
+              })
 
-    })
-    .catch(error => console.log(error));
+            })
+            .catch(error => {
+              //console.log(error)
+              res.render('index/errorPage', {
+                error: error
+              })
+            });
+        }
+      })
+      .catch(error => {
+        res.render('index/errorPage', {
+          error: error
+        })
+      });
 });
 
 // Create Vehicle Type
@@ -59,8 +84,8 @@ router.post('/',
   (req, res) => {
 
     // Set Variables
-    let max_pax = parseInt(req.body.max_pax);
-    let name = req.body.name.trim();
+    let max_pax = parseInt(req.body.max_pax).replace(/<(?:.|\n)*?>/gm, '');
+    let name = req.body.name.trim().replace(/<(?:.|\n)*?>/gm, '');
     let errors = [];
 
     if (!name.length > 0) {
@@ -92,7 +117,10 @@ router.post('/',
           res.redirect('/vehicle-types')
         })
         .catch(error => {
-          console.log(error);
+          //console.log(error);
+          res.render('index/errorPage', {
+            error: error
+          })
         })
     }
 })
@@ -103,47 +131,70 @@ router.post('/edit/:id',
   ensureAuthenticated,
   (req, res) => {
 
-    // Set Variables
-    let max_pax = parseInt(req.body.max_pax);
-    let name = req.body.name.trim();
-    let errors = [];
+    axios.get(`${keys.sassTransferServiceAPIURI}/api/VehicleTypes/${req.params.id}?access_token=${req.session.serverAccessToken}`)
+      .then(response => {
+        // Check that the resouce belongs to the user requesting it
+        if (response.data.companyId != req.session.companyID) {
+          req.flash('error_msg', 'Unauthorized');
+          res.redirect('/dashboard');
+        }
+        else {
+          // Set Variables
+          let max_pax = parseInt(req.body.max_pax).replace(/<(?:.|\n)*?>/gm, '');
+          let name = req.body.name.trim().replace(/<(?:.|\n)*?>/gm, '');
+          let errors = [];
 
-    if (!name.length > 0) {
-      errors.push({ text: 'Name cannot be blank.' })
-    }
-    if (!max_pax > 0) {
-      errors.push({ text: 'Maximum Passengers must be more than zero.' })
-    }
+          if (!name.length > 0) {
+            errors.push({ text: 'Name cannot be blank.' })
+          }
+          if (!max_pax > 0) {
+            errors.push({ text: 'Maximum Passengers must be more than zero.' })
+          }
 
-    if (errors.length > 0) {      
-      axios.get(`${keys.sassTransferServiceAPIURI}/api/VehicleTypes/${req.params.id}?&access_token=${req.session.serverAccessToken}`)
-        .then(response => {
-          res.render('dashboard', {
-            dashboardLink: true,
-            vehicleTypesEditActive: true,
-            companyID: req.session.companyID,
-            vehicleType: response.data,
-            errors
-          });
-        })
-        .catch(error => console.log(error));
-    }
-    else {
-      // Send Post Request to API Server
-      axios.put(`${keys.sassTransferServiceAPIURI}/api/VehicleTypes/${req.params.id}?&access_token=${req.session.serverAccessToken}`, {
-        max_pax,        
-        name,
-        companyId: req.session.companyID
+          if (errors.length > 0) {
+            axios.get(`${keys.sassTransferServiceAPIURI}/api/VehicleTypes/${req.params.id}?&access_token=${req.session.serverAccessToken}`)
+              .then(response => {
+                res.render('dashboard', {
+                  dashboardLink: true,
+                  vehicleTypesEditActive: true,
+                  companyID: req.session.companyID,
+                  vehicleType: response.data,
+                  errors
+                });
+              })
+              .catch(error => {
+                //console.log(error)
+                res.render('index/errorPage', {
+                  error: error
+                })
+              });
+          }
+          else {
+            // Send Post Request to API Server
+            axios.put(`${keys.sassTransferServiceAPIURI}/api/VehicleTypes/${req.params.id}?&access_token=${req.session.serverAccessToken}`, {
+              max_pax,
+              name,
+              companyId: req.session.companyID
+            })
+              .then(response => {
+                req.flash('success_msg', 'Vehicle Type updated.')
+                res.redirect('/vehicle-types')
+              })
+              .catch(error => {
+                //console.log(error);
+                res.render('index/errorPage', {
+                  error: error
+                })
+              })
+
+          }
+        }
       })
-        .then(response => {
-          req.flash('success_msg', 'Vehicle Type updated.')
-          res.redirect('/vehicle-types')
+      .catch(error => {
+        res.render('index/errorPage', {
+          error: error
         })
-        .catch(error => {
-          console.log(error);
-        })
-
-    }
+      });
   })
 
 // Delete Vehicle Type
@@ -151,19 +202,37 @@ router.get('/delete/:id',
   authenticateServer,
   ensureAuthenticated,
   (req, res) => {
-    // Send Post Request to API Server
-    axios.delete(`${keys.sassTransferServiceAPIURI}/api/VehicleTypes/${req.params.id}?&access_token=${req.session.serverAccessToken}`, {
-      body: JSON.stringify([
-        req.params.id
-      ])
-    })
+    axios.get(`${keys.sassTransferServiceAPIURI}/api/VehicleTypes/${req.params.id}?access_token=${req.session.serverAccessToken}`)
       .then(response => {
-        req.flash('success_msg', 'Vehicle Type deleted.')
-        res.redirect('/vehicle-types')
+        // Check that the resouce belongs to the user requesting it
+        if (response.data.companyId != req.session.companyID) {
+          req.flash('error_msg', 'Unauthorized');
+          res.redirect('/dashboard');
+        }
+        else {
+          // Send Post Request to API Server
+          axios.delete(`${keys.sassTransferServiceAPIURI}/api/VehicleTypes/${req.params.id}?&access_token=${req.session.serverAccessToken}`, {
+            body: JSON.stringify([
+              req.params.id
+            ])
+          })
+            .then(response => {
+              req.flash('success_msg', 'Vehicle Type deleted.')
+              res.redirect('/vehicle-types')
+            })
+            .catch(error => {
+              //console.log(error);
+              res.render('index/errorPage', {
+                error: error
+              })
+            })
+        }
       })
       .catch(error => {
-        console.log(error);
-      })
+        res.render('index/errorPage', {
+          error: error
+        })
+      });
 })
 
 module.exports = router;
